mcp-notion-email/requirements.txt
groq==0.9.0
python-dotenv==1.0.0
mcp==1.0.0
asyncio==3.4.3
aiohttp==3.9.0